New Apocalypse event animation improvements.
Might be a candidate for appearing in the "released" product.

Initial background for events is now metallic with a bit of sheen
to it.

Background remains for a bit longer now after bouncing completes.

The Shine/Shimmer layer is now brighter, to account for the new metallic
background. The Shimmer animation doesn't begin now until the
bouncing is complete, allowing it to be observed more easily.

Looks freaking great!

Borders appearing between events is now visbile on both white
and black backgrounds.